Update more-elements.vue
In the case of the v-model binding data, the camel case naming convention used here did not achieve the expected dynamic response of the data. This was verified by editing in Codepen.io. I tried to use the attribute writing in the document to get what I think For the desired effect, I hope the author will modify it.